corrigindo delete receita sem foto

+++ pwa-firebase/js/db.js
@@ -12,6 +12,7 @@ db.enablePersistence()
     }
   });
 
+
 // real-time listener, sempre tiver uma mudança no db ele manda um snapshot
 db.collection('recipes').onSnapshot(snapshot => {
   //console.log(snapshot.docChanges());
@@ -28,9 +29,12 @@ db.collection('recipes').onSnapshot(snapshot => {
       // remove a receita
       removeRecipe(change.doc.id);
       //remove foto da receita do storage
-      //console.log(firebase.storage().ref('Imagens/' + change.doc.id + ".png"))
-      firebase.storage().ref('Imagens/' + change.doc.id + ".png").delete();
-      firebase.database().ref('Imagens/'+change.doc.id).remove();
+      firebase.database().ref('Imagens/'+change.doc.id).on('value' , function(snapshot){
+        if(snapshot.exists()){
+          firebase.storage().ref('Imagens/' + change.doc.id + ".png").delete();
+          firebase.database().ref('Imagens/'+change.doc.id).remove();
+        }
+      });
     }
   });
 });
